Moved registerFetchThreadListAsync from chatsScreen to MainTabNavigator
